fix: rename JobTitle to Specialty
